# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DHD Livraison HR Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /employees:
    get:
      operationId: listEmployees
      tags: [employees]
      summary: List employees, optionally filtered by office
      parameters:
        - name: officeId
          in: query
          schema:
            type: number
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of employees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
    post:
      operationId: createEmployee
      tags: [employees]
      summary: Create a new employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeInput"
      responses:
        "201":
          description: Employee created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /employees/{id}:
    get:
      operationId: getEmployee
      tags: [employees]
      summary: Get employee by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Employee details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "404":
          description: Employee not found
    patch:
      operationId: updateEmployee
      tags: [employees]
      summary: Update employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeUpdate"
      responses:
        "200":
          description: Employee updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    delete:
      operationId: deleteEmployee
      tags: [employees]
      summary: Delete employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
  /employees/attendance-summary:
    get:
      operationId: getEmployeeAttendanceSummary
      tags: [employees]
      summary: Get attendance summary for an employee
      parameters:
        - name: employeeId
          in: query
          required: true
          schema:
            type: number
        - name: month
          in: query
          schema:
            type: string
        - name: year
          in: query
          schema:
            type: number
      responses:
        "200":
          description: Attendance summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceSummary"
  /employees/{id}/salary-history:
    get:
      operationId: getEmployeeSalaryHistory
      tags: [employees]
      summary: Get salary history for an employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Salary history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Salary"

  # ── ATTENDANCE ────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Employee:
      type: object
      required: [id, officeId, firstName, lastName, phone, position, baseSalary, workStartTime, workEndTime, createdAt]
      properties:
        id:
          type: number
        officeId:
          type: number
        officeName:
          type: ["string", "null"]
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        position:
          type: string
        baseSalary:
          type: number
        workStartTime:
          type: string
          description: "HH:MM format"
        workEndTime:
          type: string
          description: "HH:MM format"
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    EmployeeInput:
      type: object
      required: [officeId, firstName, lastName, phone, position, baseSalary, workStartTime, workEndTime]
      properties:
        officeId:
          type: number
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        email:
          type: string
        position:
          type: string
        baseSalary:
          type: number
        workStartTime:
          type: string
        workEndTime:
          type: string
    EmployeeUpdate:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        email:
          type: string
        position:
          type: string
        baseSalary:
          type: number
        workStartTime:
          type: string
        workEndTime:
          type: string
        isActive:
          type: boolean
    AttendanceSummary:
      type: object
      required: [employeeId, month, year, presentDays, absentDays, lateDays, totalHours, overtimeHours, totalDeductions, totalBonuses]
      properties:
        employeeId:
          type: number
        month:
          type: string
        year:
          type: number
        presentDays:
          type: number
        absentDays:
          type: number
        lateDays:
          type: number
        totalHours:
          type: number
        overtimeHours:
          type: number
        totalDeductions:
          type: number
        totalBonuses:
          type: number

    # ATTENDANCE
...
    Salary:
      type: object
      required: [id, employeeId, month, year, baseSalary, status, createdAt]
      properties:
        id:
          type: number
        employeeId:
          type: number
        employeeName:
          type: ["string", "null"]
        officeId:
          type: ["number", "null"]
        officeName:
          type: ["string", "null"]
        month:
          type: string
        year:
          type: number
        baseSalary:
          type: number
        presentDays:
          type: number
        absentDays:
          type: number
        workedHours:
          type: number
        overtimeHours:
          type: number
        overtimeBonus:
          type: number
        lateDeductions:
          type: number
        advanceDeductions:
          type: number
        otherDeductions:
          type: number
        violationDeductions:
          type: number
        bonuses:
          type: number
        finalSalary:
          type: number
        status:
          type: string
          enum: [pending, paid, postponed]
        paidAt:
          type: ["string", "null"]
          format: date-time
        postponedUntil:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
...